# contracts/openapi/agent-internal-api.yaml
openapi: 3.1.0
info:
  title: Freelance Ops Agent Internal API
  version: 0.4.0
  description: Spring Boot가 FastAPI/LangGraph run을 제어하는 내부 전용 계약
servers:
  - url: http://agent:8000
    description: Docker internal network
...
paths:
...
  /internal/v1/agent-runs:
    post:
      tags: [AgentRuns]
      operationId: startAgentRun
      security:
        - delegationToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunRequest"
      responses:
        "202":
          description: Run accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunAccepted"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/agent-runs/{runId}:
    get:
      tags: [AgentRuns]
      operationId: getAgentRun
      security:
        - delegationToken: []
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Current public run state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunView"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/agent-runs/{runId}/resume:
    post:
      tags: [AgentRuns]
      operationId: resumeAgentRun
      security:
        - delegationToken: []
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeAgentRunRequest"
      responses:
        "202":
          description: Resume command accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunAccepted"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/agent-runs/{runId}/events:
    get:
      tags: [AgentRuns]
      operationId: streamAgentRunEvents
      description: Streams ordered run events as server-sent events and supports reconnect cursors.
      security:
        - delegationToken: []
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Ordered Agent run event stream
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/agent-runs/{runId}/route-observations:
    get:
      tags: [AgentRuns]
      operationId: getRouteObservations
      description: Returns a finite, cursor-based snapshot of persisted route.selected events.
      security:
        - delegationToken: []
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: After-Event-ID
          in: header
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: At most 100 route observations after the supplied cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteObservationBatch"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/agent-runs/{runId}/cancel:
    post:
      tags: [AgentRuns]
      operationId: cancelAgentRun
      security:
        - delegationToken: []
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Cancelled run state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunView"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    delegationToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    metricsToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived audience-bound token issued by Spring Boot
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ProblemResponse:
      description: RFC 9457 compatible problem response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    TrustedRunContext:
      type: object
      additionalProperties: false
      required:
        - runId
        - threadId
        - traceId
        - workspaceId
        - projectId
        - initiatedBy
        - effectivePermissions
      properties:
        runId:
          type: string
          format: uuid
        threadId:
          type: string
          format: uuid
        traceId:
          type: string
          minLength: 1
          maxLength: 128
        workspaceId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        initiatedBy:
          type: string
          format: uuid
        effectivePermissions:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
    RunBudget:
      type: object
      additionalProperties: false
      required:
        - maxDurationSeconds
        - maxModelCalls
        - maxToolCalls
        - maxInputTokens
        - maxOutputTokens
        - maxDepartments
        - maxHierarchyDepth
        - maxSearchCredits
        - maxRetries
        - maxHandoffs
      properties:
        maxDurationSeconds:
          type: integer
          minimum: 1
          maximum: 900
        maxModelCalls:
          type: integer
          minimum: 0
          maximum: 50
        maxToolCalls:
          type: integer
          minimum: 0
          maximum: 100
        maxInputTokens:
          type: integer
          minimum: 0
        maxOutputTokens:
          type: integer
          minimum: 0
        maxDepartments:
          type: integer
          minimum: 1
          maximum: 4
        maxHierarchyDepth:
          type: integer
          minimum: 1
          maximum: 2
        maxSearchCredits:
          type: integer
          minimum: 0
          maximum: 100
          default: 0
        maxRetries:
          type: integer
          minimum: 0
          maximum: 5
          default: 1
        maxHandoffs:
          type: integer
          minimum: 0
          maximum: 10
          default: 3
    ModelSelection:
      type: object
      additionalProperties: false
      required: [provider, model]
      properties:
        provider:
          type: string
          enum: [OPENAI, GEMINI]
        model:
          type: string
          minLength: 1
          maxLength: 100
        reasoningEffort:
          type: string
          enum: [NONE, LOW, MEDIUM, HIGH]
          default: LOW
...
    AgentRunRequest:
      type: object
      additionalProperties: false
      required: [context, budget, modelSelection, safetyContext, input]
      properties:
        context:
          $ref: "#/components/schemas/TrustedRunContext"
        budget:
          $ref: "#/components/schemas/RunBudget"
        modelSelection:
          $ref: "#/components/schemas/ModelSelection"
        safetyContext:
          $ref: "#/components/schemas/SafetyContext"
        clarificationHistory:
          type: array
          maxItems: 30
          default: []
          items:
            $ref: "#/components/schemas/ClarificationAnswer"
        input:
          type: object
          additionalProperties: false
          required: [requirementText]
          properties:
            requirementText:
              type: string
              minLength: 1
              maxLength: 50000
            locale:
              type: string
              default: ko-KR
            jurisdictionCode:
              type: string
              minLength: 2
              maxLength: 32
            directToolOperation:
              type: string
              enum: [GET_PROJECT_CONTEXT]
            workflowMode:
              type: string
              enum: [PROJECT_ANALYSIS, AD_HOC]
              default: PROJECT_ANALYSIS
              description: Full project analysis is the default; AD_HOC is reserved for bounded internal tasks.
    ClarificationAnswer:
      type: object
      additionalProperties: false
      required: [question, answer]
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 5000
        answer:
          type: string
          minLength: 1
          maxLength: 5000
    SafetyContext:
      type: object
      additionalProperties: false
      properties:
        externalSideEffect:
          type: boolean
          default: false
        sensitiveData:
          type: boolean
          default: false
        financialAuthorityRequired:
          type: boolean
          default: false
        legalAuthorityRequired:
          type: boolean
          default: false
        irreversibleAction:
          type: boolean
          default: false
        approvalRequired:
          type: boolean
          default: false
        authorityVerified:
          type: boolean
          default: false
    AgentRunAccepted:
      type: object
      additionalProperties: false
      required: [runId, status, acceptedAt]
      properties:
        runId:
          type: string
          format: uuid
        status:
          type: string
          const: QUEUED
        acceptedAt:
          type: string
          format: date-time
    AgentRunStatus:
      type: string
      enum:
        - QUEUED
        - RUNNING
        - WAITING_FOR_USER
        - COMPLETED
        - PARTIAL
        - FAILED
        - CANCELLED
    AgentRunEvent:
      type: object
      additionalProperties: false
      required: [eventId, runId, type, occurredAt, data]
      properties:
        eventId:
          type: integer
          minimum: 1
        runId:
          type: string
          format: uuid
        type:
          type: string
          pattern: "^[a-z]+(?:\\.[a-z]+)+$"
          maxLength: 100
        occurredAt:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
    RouteObservationBatch:
      type: object
      additionalProperties: false
      required: [runId, status, events, nextEventId, hasMore, terminal]
      properties:
        runId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AgentRunStatus"
        events:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/AgentRunEvent"
        nextEventId:
          type: integer
          minimum: 0
        hasMore:
          type: boolean
        terminal:
          type: boolean
    DepartmentName:
      type: string
      enum: [REQUIREMENTS, RESEARCH, DEAL_DESIGN, VERIFICATION]
    SourceReference:
      type: object
      additionalProperties: false
      required: [title, url, provider, contentSha256, fetchedAt, authorityLevel, excerpt]
      properties:
        title:
          type: string
          maxLength: 500
        url:
          type: string
          format: uri
          maxLength: 2048
        provider:
          type: string
          maxLength: 50
        contentSha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        fetchedAt:
          type: string
          format: date-time
        authorityLevel:
          type: string
          maxLength: 50
        jurisdiction:
          type: string
          maxLength: 32
        excerpt:
          type: string
          minLength: 1
          maxLength: 4000
    DepartmentResult:
      type: object
      additionalProperties: false
      required: [department, status, summary, evidenceIds, assumptionIds, sources]
      properties:
        department:
          $ref: "#/components/schemas/DepartmentName"
        status:
          type: string
          enum: [COMPLETED, PARTIAL, FAILED, SKIPPED]
        summary:
          type: string
          maxLength: 10000
        evidenceIds:
          type: array
          items:
            type: string
            format: uuid
        assumptionIds:
          type: array
          items:
            type: string
            format: uuid
        sources:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/SourceReference"
        errorCode:
          type: string
          maxLength: 100
    AgentRunResult:
      type: object
      additionalProperties: false
      required: [projectSummary, openQuestions, departmentResults, quotationDraft, quotationDrafts]
      properties:
        projectSummary:
          type: string
          maxLength: 10000
        openQuestions:
          type: array
          items:
            type: string
            maxLength: 1000
        departmentResults:
          type: array
          maxItems: 4
          items:
            $ref: "#/components/schemas/DepartmentResult"
        quotationDraft:
          oneOf:
            - $ref: "#/components/schemas/QuotationDraft"
            - type: "null"
          description: Deprecated compatibility view of the RECOMMENDED draft.
        quotationDrafts:
          type: array
          maxItems: 3
          description: AI-generated LEAN, RECOMMENDED, and EXPANDED drafts with distinct scope and effort.
          items:
            $ref: "#/components/schemas/QuotationDraft"
    QuotationDraft:
      type: object
      additionalProperties: false
      required: [scenario, items]
      properties:
        scenario:
          type: string
          enum: [LEAN, RECOMMENDED, EXPANDED]
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/QuotationDraftItem"
    QuotationDraftItem:
      type: object
      additionalProperties: false
      required: [title, description, quantity, unit, rateCardHint, basis]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          type: string
          enum: [HOUR, DAY, FIXED]
        rateCardHint:
          type: [string, "null"]
          maxLength: 200
        basis:
          $ref: "#/components/schemas/QuotationDraftBasis"
    QuotationDraftBasis:
      type: object
      additionalProperties: false
      required: [type, content, sourceReference, sourceTitle]
      properties:
        type:
          type: string
          enum: [ASSUMPTION, EVIDENCE]
        content:
          type: string
          minLength: 1
          maxLength: 3000
        sourceReference:
          type: [string, "null"]
          maxLength: 2048
        sourceTitle:
          type: [string, "null"]
          maxLength: 300
    AgentRunView:
      type: object
      additionalProperties: false
      required: [runId, status, metadata, updatedAt]
      properties:
        runId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AgentRunStatus"
        activeDepartment:
          $ref: "#/components/schemas/DepartmentName"
        interruption:
          $ref: "#/components/schemas/AgentInterruption"
        result:
          $ref: "#/components/schemas/AgentRunResult"
        errorCode:
          type: string
          maxLength: 100
        metadata:
          $ref: "#/components/schemas/AgentRunMetadata"
        usage:
          $ref: "#/components/schemas/AgentRunUsage"
        updatedAt:
          type: string
          format: date-time
    AgentRunMetadata:
      type: object
      additionalProperties: false
      required: [provider, model, promptVersion, toolSchemaVersion, traceId]
      properties:
        provider:
          type: string
          enum: [OPENAI, GEMINI]
        model:
          type: string
          minLength: 1
          maxLength: 100
        promptVersion:
          type: string
          minLength: 1
          maxLength: 100
        toolSchemaVersion:
          type: string
          minLength: 1
          maxLength: 100
        traceId:
          type: string
          minLength: 1
          maxLength: 128
    AgentRunUsage:
      type: object
      additionalProperties: false
      required:
        - requestTier
        - modelCalls
        - toolCalls
        - inputTokens
        - outputTokens
        - cachedTokens
        - searchCredits
        - crawledPages
        - retryCount
        - durationMs
      properties:
        requestTier:
          type: string
          enum: [DIRECT_TOOL, SINGLE_AGENT, DEPARTMENT, MULTI_DEPARTMENT, HUMAN_REQUIRED]
        modelCalls: {type: integer, minimum: 0}
        toolCalls: {type: integer, minimum: 0}
        inputTokens: {type: integer, minimum: 0}
        outputTokens: {type: integer, minimum: 0}
        cachedTokens: {type: integer, minimum: 0}
        searchCredits: {type: integer, minimum: 0}
        crawledPages: {type: integer, minimum: 0}
        retryCount: {type: integer, minimum: 0}
        durationMs: {type: integer, minimum: 0}
    AgentInterruption:
      type: object
      additionalProperties: false
      required: [interruptionId, kind, questions]
      properties:
        interruptionId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [CLARIFICATION, RISK_DECISION, QUOTE_APPROVAL]
        questions:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: string
            maxLength: 1000
    ResumeAgentRunRequest:
      type: object
      additionalProperties: false
      required: [interruptionId, idempotencyKey, answers]
      properties:
        interruptionId:
          type: string
          format: uuid
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
        answers:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [questionIndex, answer]
            properties:
              questionIndex:
                type: integer
                minimum: 0
              answer:
                type: string
                minLength: 1
                maxLength: 5000
...
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string